+++ History.md
@@ -1,4 +1,19 @@
 
+0.2.0 / 2013-02-23 
+==================
+
+  * add event delegation example
+  * add manipulation example
+  * add initial .remove() implementation
+  * add .text(str) tmp implementation
+  * add `.prop()` method
+  * add `.get()` == `.get(0)`
+  * add `.appendTo()`
+  * add `.removeClass(regexp)` support
+  * add `.css(object)` support
+  * change .append() and .prepend() to return new list
+  * change `.attr()` to set attribute for each element
+
 0.1.0 / 2013-01-04
 ==================